pyproject.toml
[project]
name = "oraculo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.4.0",
    "fake-useragent>=2.2.0",
    "langchain>=1.2.0",
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.6",
    "langchain-openai>=1.1.6",
    "python-dotenv>=1.0.0",
    "langgraph>=1.0.5",
    "playwright>=1.57.0",
    "streamlit>=1.52.2",
    "tiktoken>=0.12.0",
    "unstructured>=0.18.24",
]
